FINERACT-2169: Refactor account transfer api controller;
