+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/account/api/AccountTransfersApiResource.java
@@ -26,6 +26,7 @@ import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.ws.rs.BeanParam;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
@@ -33,22 +34,19 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.QueryParam;
-import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
-import jakarta.ws.rs.core.UriInfo;
 import lombok.RequiredArgsConstructor;
 import org.apache.fineract.commands.domain.CommandWrapper;
 import org.apache.fineract.commands.service.CommandWrapperBuilder;
 import org.apache.fineract.commands.service.PortfolioCommandSourceWritePlatformService;
-import org.apache.fineract.infrastructure.core.api.ApiRequestParameterHelper;
 import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
-import org.apache.fineract.infrastructure.core.serialization.ApiRequestJsonSerializationSettings;
 import org.apache.fineract.infrastructure.core.serialization.DefaultToApiJsonSerializer;
 import org.apache.fineract.infrastructure.core.service.Page;
 import org.apache.fineract.infrastructure.core.service.SearchParameters;
 import org.apache.fineract.infrastructure.security.service.PlatformSecurityContext;
-import org.apache.fineract.infrastructure.security.service.SqlValidator;
 import org.apache.fineract.portfolio.account.data.AccountTransferData;
+import org.apache.fineract.portfolio.account.data.request.AccountTransSearchParam;
+import org.apache.fineract.portfolio.account.data.request.AccountTransferRequest;
 import org.apache.fineract.portfolio.account.service.AccountTransfersReadPlatformService;
 import org.springframework.stereotype.Component;
 
@@ -61,9 +59,7 @@ public class AccountTransfersApiResource {
     private final PlatformSecurityContext context;
     private final DefaultToApiJsonSerializer<AccountTransferData> toApiJsonSerializer;
     private final PortfolioCommandSourceWritePlatformService commandsSourceWritePlatformService;
-    private final ApiRequestParameterHelper apiRequestParameterHelper;
     private final AccountTransfersReadPlatformService accountTransfersReadPlatformService;
-    private final SqlValidator sqlValidator;
 
     @GET
     @Path("template")
@@ -76,38 +72,28 @@ public class AccountTransfersApiResource {
             + "accounttransfers/template?fromClientId=1&fromAccountType=2&fromAccountId=1")
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "OK", content = @Content(schema = @Schema(implementation = AccountTransfersApiResourceSwagger.GetAccountTransfersTemplateResponse.class))) })
-    public String template(@QueryParam("fromOfficeId") @Parameter(description = "fromOfficeId") final Long fromOfficeId,
-            @QueryParam("fromClientId") @Parameter(description = "fromClientId") final Long fromClientId,
-            @QueryParam("fromAccountId") @Parameter(description = "fromAccountId") final Long fromAccountId,
-            @QueryParam("fromAccountType") @Parameter(description = "fromAccountType") final Integer fromAccountType,
-            @QueryParam("toOfficeId") @Parameter(description = "toOfficeId") final Long toOfficeId,
-            @QueryParam("toClientId") @Parameter(description = "toClientId") final Long toClientId,
-            @QueryParam("toAccountId") @Parameter(description = "toAccountId") final Long toAccountId,
-            @QueryParam("toAccountType") @Parameter(description = "toAccountType") final Integer toAccountType,
-            @Context final UriInfo uriInfo) {
+    public AccountTransferData template(@BeanParam AccountTransSearchParam accountTransSearchParam) {
 
-        this.context.authenticatedUser().validateHasReadPermission(AccountTransfersApiConstants.ACCOUNT_TRANSFER_RESOURCE_NAME);
+        context.authenticatedUser().validateHasReadPermission(AccountTransfersApiConstants.ACCOUNT_TRANSFER_RESOURCE_NAME);
 
-        final AccountTransferData transferData = this.accountTransfersReadPlatformService.retrieveTemplate(fromOfficeId, fromClientId,
-                fromAccountId, fromAccountType, toOfficeId, toClientId, toAccountId, toAccountType);
-
-        final ApiRequestJsonSerializationSettings settings = this.apiRequestParameterHelper.process(uriInfo.getQueryParameters());
-        return this.toApiJsonSerializer.serialize(settings, transferData, AccountTransfersApiConstants.RESPONSE_DATA_PARAMETERS);
+        return accountTransfersReadPlatformService.retrieveTemplate(accountTransSearchParam.getFromAccountId(),
+                accountTransSearchParam.getFromClientId(), accountTransSearchParam.getFromAccountId(),
+                accountTransSearchParam.getFromAccountType(), accountTransSearchParam.getToOfficeId(),
+                accountTransSearchParam.getToClientId(), accountTransSearchParam.getToAccountId(),
+                accountTransSearchParam.getToAccountType());
     }
 
     @POST
     @Consumes({ MediaType.APPLICATION_JSON })
     @Produces({ MediaType.APPLICATION_JSON })
     @Operation(summary = "Create new Transfer", description = "Ability to create new transfer of monetary funds from one account to another.")
-    @RequestBody(required = true, content = @Content(schema = @Schema(implementation = AccountTransfersApiResourceSwagger.PostAccountTransfersRequest.class)))
+    @RequestBody(required = true, content = @Content(schema = @Schema(implementation = AccountTransferRequest.class)))
     @ApiResponse(responseCode = "200", description = "OK", content = @Content(schema = @Schema(implementation = AccountTransfersApiResourceSwagger.PostAccountTransfersResponse.class)))
-    public String create(@Parameter(hidden = true) final String apiRequestBodyAsJson) {
+    public CommandProcessingResult create(@Parameter(hidden = true) AccountTransferRequest accountTransferRequest) {
+        final CommandWrapper commandRequest = new CommandWrapperBuilder().createAccountTransfer()
+                .withJson(toApiJsonSerializer.serialize(accountTransferRequest)).build();
 
-        final CommandWrapper commandRequest = new CommandWrapperBuilder().createAccountTransfer().withJson(apiRequestBodyAsJson).build();
-
-        final CommandProcessingResult result = this.commandsSourceWritePlatformService.logCommandSource(commandRequest);
-
-        return this.toApiJsonSerializer.serialize(result);
+        return commandsSourceWritePlatformService.logCommandSource(commandRequest);
     }
 
     @GET
@@ -117,26 +103,19 @@ public class AccountTransfersApiResource {
             + "accounttransfers")
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "OK", content = @Content(schema = @Schema(implementation = AccountTransfersApiResourceSwagger.GetAccountTransfersResponse.class))) })
-    public String retrieveAll(@Context final UriInfo uriInfo,
-            @QueryParam("externalId") @Parameter(description = "externalId") final String externalId,
+    public Page<AccountTransferData> retrieveAll(@QueryParam("externalId") @Parameter(description = "externalId") final String externalId,
             @QueryParam("offset") @Parameter(description = "offset") final Integer offset,
             @QueryParam("limit") @Parameter(example = "limit") final Integer limit,
             @QueryParam("orderBy") @Parameter(description = "orderBy") final String orderBy,
             @QueryParam("sortOrder") @Parameter(description = "sortOrder") final String sortOrder,
             @QueryParam("accountDetailId") @Parameter(description = "accountDetailId") final Long accountDetailId) {
 
-        this.context.authenticatedUser().validateHasReadPermission(AccountTransfersApiConstants.ACCOUNT_TRANSFER_RESOURCE_NAME);
+        context.authenticatedUser().validateHasReadPermission(AccountTransfersApiConstants.ACCOUNT_TRANSFER_RESOURCE_NAME);
 
-        sqlValidator.validate(orderBy);
-        sqlValidator.validate(sortOrder);
-        sqlValidator.validate(externalId);
         final SearchParameters searchParameters = SearchParameters.builder().limit(limit).externalId(externalId).offset(offset)
                 .orderBy(orderBy).sortOrder(sortOrder).build();
 
-        final Page<AccountTransferData> transfers = this.accountTransfersReadPlatformService.retrieveAll(searchParameters, accountDetailId);
-
-        final ApiRequestJsonSerializationSettings settings = this.apiRequestParameterHelper.process(uriInfo.getQueryParameters());
-        return this.toApiJsonSerializer.serialize(settings, transfers, AccountTransfersApiConstants.RESPONSE_DATA_PARAMETERS);
+        return accountTransfersReadPlatformService.retrieveAll(searchParameters, accountDetailId);
     }
 
     @GET
@@ -147,15 +126,9 @@ public class AccountTransfersApiResource {
             + "accounttransfers/1")
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "OK", content = @Content(schema = @Schema(implementation = AccountTransfersApiResourceSwagger.GetAccountTransfersResponse.GetAccountTransfersPageItems.class))) })
-    public String retrieveOne(@PathParam("transferId") @Parameter(description = "transferId") final Long transferId,
-            @Context final UriInfo uriInfo) {
-
-        this.context.authenticatedUser().validateHasReadPermission(AccountTransfersApiConstants.ACCOUNT_TRANSFER_RESOURCE_NAME);
-
-        final AccountTransferData transfer = this.accountTransfersReadPlatformService.retrieveOne(transferId);
-
-        final ApiRequestJsonSerializationSettings settings = this.apiRequestParameterHelper.process(uriInfo.getQueryParameters());
-        return this.toApiJsonSerializer.serialize(settings, transfer, AccountTransfersApiConstants.RESPONSE_DATA_PARAMETERS);
+    public AccountTransferData retrieveOne(@PathParam("transferId") @Parameter(description = "transferId") final Long transferId) {
+        context.authenticatedUser().validateHasReadPermission(AccountTransfersApiConstants.ACCOUNT_TRANSFER_RESOURCE_NAME);
+        return accountTransfersReadPlatformService.retrieveOne(transferId);
     }
 
     @GET
@@ -167,23 +140,13 @@ public class AccountTransfersApiResource {
             + "accounttransfers/templateRefundByTransfer?fromAccountId=2&fromAccountType=1& toAccountId=1&toAccountType=2&toClientId=1&toOfficeId=1")
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "OK", content = @Content(schema = @Schema(implementation = AccountTransfersApiResourceSwagger.GetAccountTransfersTemplateRefundByTransferResponse.class))) })
-    public String templateRefundByTransfer(@QueryParam("fromOfficeId") @Parameter(description = "fromOfficeId") final Long fromOfficeId,
-            @QueryParam("fromClientId") @Parameter(description = "fromClientId") final Long fromClientId,
-            @QueryParam("fromAccountId") @Parameter(description = "fromAccountId") final Long fromAccountId,
-            @QueryParam("fromAccountType") @Parameter(description = "fromAccountType") final Integer fromAccountType,
-            @QueryParam("toOfficeId") @Parameter(description = "toOfficeId") final Long toOfficeId,
-            @QueryParam("toClientId") @Parameter(description = "toClientId") final Long toClientId,
-            @QueryParam("toAccountId") @Parameter(description = "toAccountId") final Long toAccountId,
-            @QueryParam("toAccountType") @Parameter(description = "toAccountType") final Integer toAccountType,
-            @Context final UriInfo uriInfo) {
-
-        this.context.authenticatedUser().validateHasReadPermission(AccountTransfersApiConstants.ACCOUNT_TRANSFER_RESOURCE_NAME);
-
-        final AccountTransferData transferData = this.accountTransfersReadPlatformService.retrieveRefundByTransferTemplate(fromOfficeId,
-                fromClientId, fromAccountId, fromAccountType, toOfficeId, toClientId, toAccountId, toAccountType);
-
-        final ApiRequestJsonSerializationSettings settings = this.apiRequestParameterHelper.process(uriInfo.getQueryParameters());
-        return this.toApiJsonSerializer.serialize(settings, transferData, AccountTransfersApiConstants.RESPONSE_DATA_PARAMETERS);
+    public AccountTransferData templateRefundByTransfer(@BeanParam AccountTransSearchParam accountTransSearchParam) {
+        context.authenticatedUser().validateHasReadPermission(AccountTransfersApiConstants.ACCOUNT_TRANSFER_RESOURCE_NAME);
+        return accountTransfersReadPlatformService.retrieveRefundByTransferTemplate(accountTransSearchParam.getFromAccountId(),
+                accountTransSearchParam.getFromClientId(), accountTransSearchParam.getFromAccountId(),
+                accountTransSearchParam.getFromAccountType(), accountTransSearchParam.getToOfficeId(),
+                accountTransSearchParam.getToClientId(), accountTransSearchParam.getToAccountId(),
+                accountTransSearchParam.getToAccountType());
     }
 
     @POST
@@ -191,15 +154,12 @@ public class AccountTransfersApiResource {
     @Consumes({ MediaType.APPLICATION_JSON })
     @Produces({ MediaType.APPLICATION_JSON })
     @Operation(summary = "Refund of an Active Loan by Transfer", description = "Ability to refund an active loan by transferring to a savings account.")
-    @RequestBody(required = true, content = @Content(schema = @Schema(implementation = AccountTransfersApiResourceSwagger.PostAccountTransfersRefundByTransferRequest.class)))
+    @RequestBody(required = true, content = @Content(schema = @Schema(implementation = AccountTransferRequest.class)))
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "OK", content = @Content(schema = @Schema(implementation = AccountTransfersApiResourceSwagger.PostAccountTransfersRefundByTransferResponse.class))) })
-    public String templateRefundByTransferPost(@Parameter(hidden = true) final String apiRequestBodyAsJson) {
-
-        final CommandWrapper commandRequest = new CommandWrapperBuilder().refundByTransfer().withJson(apiRequestBodyAsJson).build();
-
-        final CommandProcessingResult result = this.commandsSourceWritePlatformService.logCommandSource(commandRequest);
-
-        return this.toApiJsonSerializer.serialize(result);
+    public CommandProcessingResult templateRefundByTransferPost(@Parameter(hidden = true) AccountTransferRequest accountTransferRequest) {
+        final CommandWrapper commandRequest = new CommandWrapperBuilder().refundByTransfer()
+                .withJson(toApiJsonSerializer.serialize(accountTransferRequest)).build();
+        return commandsSourceWritePlatformService.logCommandSource(commandRequest);
     }
 }
